Tried implemeting Die() and failed haha
I created a method Die() but unfortunately it's not working.
The way I was trying to make it work is by disabling the component PlayerMovementController, but even though it's working (it is disabling) somehow the controllers still work, the script is still having an effect on the game. Destroying the component worked, but I was trying to avoid that.
Need to leave this for now unfortunately. Cheers!